Removed default run in kv_store and revised code

+++ examples/dashboard/src/util.rs
@@ -7,5 +7,5 @@ pub const OUTGOING_MESSAGE_PERIOD: Duration = Duration::from_millis(1);
 pub const WAIT_LEADER_TIMEOUT: Duration = Duration::from_millis(500);
 pub const WAIT_DECIDED_TIMEOUT: Duration = Duration::from_millis(50);
 pub const UI_TICK_PERIOD: Duration = Duration::from_millis(200);
-pub const BATCH_SIZE: usize = 100;
+pub const BATCH_SIZE: u64 = 100;
 pub const BATCH_PERIOD: Duration = Duration::from_millis(50);
